pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reelify"
version = "0.1.0"
description = "Turn long screen recordings into concise summary videos."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Reelify Authors"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "opencv-python",
    "scenedetect[opencv]",
    "typer[all]",
    "ffmpeg-python",
    "openai>=1.0",
    "google-genai>=1.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
subtitles = [
    "openai-whisper",
]
dev = [
    "pytest",
    "pytest-cov",
]

[project.scripts]
reelify = "reelify.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
